Describe operations and events WebTransport is expected to implement
Currently, in the W3C API draft, we cite the HTTP/3 draft and RFC 9000
directly.  This does not work with WebTransport over HTTP/2, hence we
need for the overview draft to provide a common set of operations that
both H3 and H2 versions of the draft are expected to implement.